refactor: correct path for import and change name folders and files to correct warnings
