+++ src/Interfaces/common.ts
