+++ src/Pages/Main/button-main-link.ts
@@ -1,5 +1,5 @@
 import BaseComponent from '../../Abstract/base-component';
-import { ILinkButtonProps } from '../../Interfaces/interfaces';
+import { ILinkButtonProps } from '../../Interfaces/common';
 import Services from '../../Interfaces/services';
 
 export default class ButtonLink extends BaseComponent {
