+++ src/Pages/Sprint/game-field.ts
@@ -1,7 +1,7 @@
 /* eslint-disable prefer-destructuring */
 import Services from '../../Interfaces/services';
 import BaseComponent from '../../Abstract/base-component';
-import ButtonWithCallback from '../../components/button-with-callback';
+import ButtonWithCallback from '../../Components/button-component';
 import AudioBtnGameSprint from './audio-btn';
 
 export default class FieldGame extends BaseComponent {
