+++ src/Pages/Sprint/game-header.ts
@@ -1,7 +1,7 @@
 import Services from '../../Interfaces/services';
 import BaseComponent from '../../Abstract/base-component';
 import Timer from './game-timer';
-import ButtonWithCallback from '../../components/button-with-callback';
+import ButtonWithCallback from '../../Components/button-component';
 
 export default class HeaderGame extends BaseComponent {
   private count?: BaseComponent;
