+++ src/Pages/Textbook/textbook.ts
@@ -1,10 +1,10 @@
 import { Word } from '../../Interfaces/word-model';
 import BaseComponent from '../../Abstract/base-component';
-import { getWords } from '../../model/getTextbook';
+import { getWords } from '../../Model/getTextbook';
 import Services from '../../Interfaces/services';
 import TextBookCart from '../../components/textbook-cart';
 import TextBookSettings from '../../components/textbook-settings';
-import State from '../../model/state';
+import State from '../../Model/state';
 
 export default class TextbookPage extends BaseComponent {
   constructor(private readonly parent: HTMLElement, private readonly services: Services) {
