+++ src/components/button-audio-textbook.ts
@@ -1,6 +1,6 @@
 import BaseComponent from '../Abstract/base-component';
-import { baseUrl } from '../model/getTextbook';
-import State from '../model/state';
+import { baseUrl } from '../Model/getTextbook';
+import State from '../Model/state';
 import Services from '../Interfaces/services';
 
 export default class ButtonAudioTextbook extends BaseComponent<HTMLButtonElement> {
