+++ src/components/container-exit-autorise.ts
@@ -1,6 +1,6 @@
 import BaseComponent from '../Abstract/base-component';
 import Services from '../Interfaces/services';
-import ButtonWithCallback from './button-with-callback';
+import ButtonWithCallback from './button-component';
 import NameUser from './name-user';
 
 export default class ContainerExitAutorise extends BaseComponent {
